Add initial map generation for IndependentBoard
Some functionalities for map generation are added.
Iterates through non-rotated cases and adds them to configuration list.
Added enums and board state, as well as changing the way height and width are obtained.

Square is moved to the top layer, and is no longer a subclass for Ship. This allows Board to use Sqaure as a coordinate storage.

 - Fixed typo in license

Signed-off-by: Shen Yichen <2007.yichen@gmail.com>