Added HIT function
Clicking on a square cycles through OPEN > MISS > HIT for the square. When a sqaure is marked HIT, it is treated as OPEN for calculation of configs and data matrices. However, on HighSeas it will be uncoloured and marked with an X.

One such example showing this special case will be to surround a square with HITs on a standard battleships game. One will see that the surrounded sqaure is still a possible position, but using MISSes instead will cause it to be uncoloured, denoting 0 possible configs exists there.

Signed-off-by: Shen Yichen <2007.yichen@gmail.com>