Corrected HIT square
HIT square will no longer be considered for max, and the possible configs is forced to be 0. As this is done in the controller, the data of the surrounding sqaures will not be affected, as it should be.

When a square is not open, the status bar shows it's status instead of the number of possible configs, which would have been zero anyway.

Signed-off-by: Shen Yichen <2007.yichen@gmail.com>